Extract CheckURLUniqueness operation to shared operations
