+++ app/services/learning/articles/operations/container.rb
@@ -7,7 +7,6 @@ module Learning
         extend Dry::Container::Mixin
         import Shared::Container
 
-        register('check_url_uniqueness') { CheckURLUniqueness.new }
         register('assign_attributes') { AssignAttributes.new }
       end
     end
