+++ app/services/learning/articles/transactions/create_transaction.rb
@@ -6,7 +6,7 @@ module Learning
       class CreateTransaction < BaseTransaction
         step :validate, with: 'shared.validate'
         step :build, with: 'shared.build'
-        step :check_url_uniqueness, with: 'check_url_uniqueness'
+        step :check_url_uniqueness, with: 'shared.check_url_uniqueness'
         step :assign_attributes, with: 'assign_attributes'
         step :persist, with: 'shared.persist'
       end
