+++ app/services/learning/articles/transactions/update_transaction.rb
@@ -5,7 +5,7 @@ module Learning
     module Transactions
       class UpdateTransaction < BaseTransaction
         step :validate, with: 'shared.validate'
-        step :check_url_uniqueness, with: 'check_url_uniqueness'
+        step :check_url_uniqueness, with: 'shared.check_url_uniqueness'
         step :assign_attributes, with: 'assign_attributes'
         step :persist, with: 'shared.persist'
       end
