+++ app/services/shared/container.rb
@@ -5,6 +5,7 @@ module Shared
     register('validate') { Operations::Validate.new }
     register('build') { Operations::Build.new }
     register('must_be_created_once_per_day') { Operations::MustBeCreatedOncePerDay.new }
+    register('check_url_uniqueness') { Operations::CheckURLUniqueness.new }
     register('persist') { Operations::Persist.new }
   end
 end
