+++ spec/services/learning/articles/transactions/update_transaction_spec.rb
@@ -9,7 +9,7 @@ RSpec.describe Learning::Articles::Transactions::UpdateTransaction do
   let(:expected_steps) do
     {
       validate: 'shared.validate',
-      check_url_uniqueness: 'check_url_uniqueness',
+      check_url_uniqueness: 'shared.check_url_uniqueness',
       assign_attributes: 'assign_attributes',
       persist: 'shared.persist'
     }
